chore: move helper
