+++ vela/repo_test.go
@@ -12,6 +12,7 @@ import (
 
 	"github.com/go-vela/server/mock/server"
 	"github.com/go-vela/types/library"
+	"github.com/go-vela/types/library/actions"
 
 	"github.com/gin-gonic/gin"
 )
@@ -450,4 +451,31 @@ func ExampleRepoService_Chown() {
 	}
 
 	fmt.Printf("Received response code %d, for repo %+v", resp.StatusCode, repo)
+}
+
+func testEvents() *library.Events {
+	return &library.Events{
+		Push: &actions.Push{
+			Branch:       Bool(true),
+			Tag:          Bool(true),
+			DeleteBranch: Bool(true),
+			DeleteTag:    Bool(true),
+		},
+		PullRequest: &actions.Pull{
+			Opened:      Bool(true),
+			Edited:      Bool(true),
+			Synchronize: Bool(true),
+			Reopened:    Bool(true),
+		},
+		Deployment: &actions.Deploy{
+			Created: Bool(true),
+		},
+		Comment: &actions.Comment{
+			Created: Bool(true),
+			Edited:  Bool(true),
+		},
+		Schedule: &actions.Schedule{
+			Run: Bool(true),
+		},
+	}
 }
